+++ internal/entity/password.go
@@ -4,7 +4,7 @@ import "time"
 
 type Password struct {
 	ID             int32      `json:"id"`
-	UserID         int32      `json:"userId"`
+	UserID         int64      `json:"userId"`
 	PasswordHash   string     `json:"passwordHash"`
 	FailedAttempts int32      `json:"failedAttempts"`
 	CreatedAt      time.Time  `json:"createdAt"`
